+++ modules/models/src/main/scala/com/odenzo/etrade/models/Account.scala
@@ -3,9 +3,6 @@ package com.odenzo.etrade.models
 import io.circe.generic.*
 import io.circe.*
 
-/** FIXME once calls working: https://apisb.etrade.com/docs/api/account/api-account-v1.html#/definitions/Account */
-case class AccountListResponse(Account: Vector[Json]) derives Codec.AsObject
-
 /**
   * @param intNo
   * @param accountId
