+++ modules/models/src/main/scala/com/odenzo/etrade/models/Message.scala
@@ -1,7 +1,6 @@
 package com.odenzo.etrade.models
 
 import com.odenzo.etrade.base.CirceUtils
-import com.odenzo.etrade.models.responses.TransactionListRs
 import io.circe.Codec
 import io.circe.generic.semiauto.deriveCodec
