+++ modules/models/src/main/scala/com/odenzo/etrade/models/Transaction.scala
@@ -18,34 +18,3 @@ case class Transaction(
     detailsURI: Option[String], // Have seen one of thee yet
     brokerage: Brokerage
 ) derives Codec.AsObject
-
-object Transaction {
-
-  val testData = """  {
-           "transactionId" : "123",
-           "accountId" : "123",
-           "transactionDate" : 123,
-           "postDate" : 1548662400000,
-           "amount" : 0.03,
-           "description" : "EXTENDED INSURANCE SWEEP DEPOSIT ACCOUNT INTEREST",
-           "transactionType" : "Interest",
-           "memo" : "",
-           "imageFlag" : false,
-           "instType" : "BROKERAGE",
-           "storeId" : 3,
-           "brokerage" : {
-             "product" : {
-               "symbol" : "#2145605",
-               "securityType" : "BOND"
-             },
-             "quantity" : 0,
-             "price" : 0,
-             "settlementCurrency" : "USD",
-             "paymentCurrency" : "USD",
-             "fee" : 0,
-             "displaySymbol" : "#2145605",
-             "settlementDate" : 1548662400000
-           },
-           "detailsURI" : "https://api.etrade.com/v1/accounts/cwrsjbz234CmJsrSi0X2T4gyA/transactions/234?storeId=3"
-         }"""
-}
